requirements.txt
browser-use>=0.1.37
langchain>=0.3.14
langchain-google-genai>=0.0.8
python-dotenv>=1.0.0
playwright>=1.40.0 